move admin dashboard pages into an admin dashboard route and layout
