+++ client/src/pages/admin-dashboard/test/competition/create.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
+import DashboardLayout from "@/components/dashboard-layout";
 import { Button } from "@/components/ui/button";
 import {
   Form,
@@ -14,19 +15,20 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { QuestionBlockManager } from "@/components/ui/Test/question-block-manager";
-import { SelectHour, SelectMinute } from "@/components/ui/Test/select-time";
 import { Textarea } from "@/components/ui/textarea";
-import { createPracticeSchema } from "@/types/practice";
+import { DateTimePicker } from "@/components/ui/time-picker/date-time-picker";
+import { NextPageWithLayout } from "@/pages/_app";
+import { createCompetitionSchema } from "@/types/competition";
 
-type Practice = z.infer<typeof createPracticeSchema>;
+type Competition = z.infer<typeof createCompetitionSchema>;
 
-export default function Create() {
-  const form = useForm<Practice>({
-    resolver: zodResolver(createPracticeSchema),
-    defaultValues: {} as Practice,
+const CreatePage: NextPageWithLayout = () => {
+  const form = useForm<Competition>({
+    resolver: zodResolver(createCompetitionSchema),
+    defaultValues: {} as Competition,
   });
 
-  const onSubmit = (data: Practice) => {
+  const onSubmit = (data: Competition) => {
     console.log("Form Data:", data);
     alert("see console for full data.");
   };
@@ -36,12 +38,11 @@ export default function Create() {
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)}>
         <h1 className="my-4 text-center text-xl font-bold">
-          Create Practice Test
+          Create Competition
         </h1>
 
         <div className="mx-auto max-w-3xl space-y-6 rounded-lg bg-gray-50 p-4 shadow-lg">
           <h3 className="-mb-2 text-lg">Basic</h3>
-
           {/* Test Name */}
           <FormField
             name="name"
@@ -57,7 +58,6 @@ export default function Create() {
               </FormItem>
             )}
           />
-
           {/* General Instructions */}
           <FormField
             name="general_instructions"
@@ -75,48 +75,38 @@ export default function Create() {
               </FormItem>
             )}
           />
-
-          {/* Time Limitation */}
-          <div>
-            <FormLabel>Time Limitation {requiredStar}</FormLabel>
-            <div className="mt-2 flex gap-4">
-              {/* Hours */}
-              <FormField
-                name="hours"
-                control={form.control}
-                render={({ field }) => (
-                  <FormItem>
-                    <FormControl>
-                      <SelectHour
-                        selectedTime={field.value}
-                        onChange={field.onChange}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              {/* Minutes */}
-              <FormField
-                name="minutes"
-                control={form.control}
-                render={({ field }) => (
-                  <FormItem>
-                    <FormControl>
-                      <SelectMinute
-                        selectedTime={field.value}
-                        onChange={field.onChange}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
+          <div className="flex">
+            {/* Starting Time */}
+            <FormField
+              name="start_time"
+              control={form.control}
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Starting Time {requiredStar}</FormLabel>
+                  <FormControl>
+                    <DateTimePicker field={field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            {/* Ending Time */}
+            <FormField
+              name="end_time"
+              control={form.control}
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Ending Time {requiredStar}</FormLabel>
+                  <FormControl>
+                    <DateTimePicker field={field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
           </div>
         </div>
-
+        {/* Question Blocks */}
         <div className="mx-auto my-4 max-w-3xl space-y-4 rounded-lg bg-gray-50 p-4 shadow-lg">
           <h3 className="-mb-2 text-lg">Question Blocks {requiredStar}</h3>
           <span className="text-xs text-gray-400">
@@ -124,11 +114,16 @@ export default function Create() {
           </span>
           <QuestionBlockManager formControl={form.control} />
         </div>
-
         <div className="flex justify-center gap-4">
           <Button type="submit">Save</Button>
         </div>
       </form>
     </Form>
   );
-}
+};
+
+CreatePage.getLayout = function getLayout(page) {
+  return <DashboardLayout>{page}</DashboardLayout>;
+};
+
+export default CreatePage;
