+++ client/src/pages/admin-dashboard/users/create/staffs.tsx
@@ -1,0 +1,16 @@
+import DashboardLayout from "@/components/dashboard-layout";
+import { NextPageWithLayout } from "@/pages/_app";
+
+const CreatePage: NextPageWithLayout = () => {
+  return (
+    <div>
+      <h1>Staffs</h1>
+    </div>
+  );
+};
+
+CreatePage.getLayout = function getLayout(page) {
+  return <DashboardLayout>{page}</DashboardLayout>;
+};
+
+export default CreatePage;
