+++ client/src/pages/admin-dashboard/users/create/teachers.tsx
@@ -1,0 +1,17 @@
+import DashboardLayout from "@/components/dashboard-layout";
+import { TeacherDataTableForm } from "@/components/ui/Users/teacher-data-table-form";
+import { NextPageWithLayout } from "@/pages/_app";
+
+const CreatePage: NextPageWithLayout = () => {
+  return (
+    <div>
+      <TeacherDataTableForm />
+    </div>
+  );
+};
+
+CreatePage.getLayout = function getLayout(page) {
+  return <DashboardLayout>{page}</DashboardLayout>;
+};
+
+export default CreatePage;
