--- client/src/pages/users/create/staffs.tsx
@@ -1,7 +1,0 @@
-export default function Create() {
-  return (
-    <div>
-      <h1>Staffs</h1>
-    </div>
-  );
-}
