--- client/src/pages/users/create/students.tsx
@@ -1,10 +1,0 @@
-import { StudentDataTableForm } from "@/components/ui/Users/student-data-table-form";
-
-export default function Create() {
-  return (
-    <div>
-      <h1 className="py-4 text-center text-2xl">Create Students</h1>
-      <StudentDataTableForm />
-    </div>
-  );
-}
