--- client/src/pages/users/create/teachers.tsx
@@ -1,9 +1,0 @@
-import { TeacherDataTableForm } from "@/components/ui/Users/teacher-data-table-form";
-
-export default function Create() {
-  return (
-    <div>
-      <TeacherDataTableForm />
-    </div>
-  );
-}
