--- client/src/pages/users/create_school.tsx
@@ -1,5 +1,0 @@
-import { SchoolDataTableForm } from "@/components/ui/Users/school-data-table-form";
-
-export default function CreateSchool() {
-  return <SchoolDataTableForm />;
-}
