--- client/src/pages/users/create_team.tsx
@@ -1,5 +1,0 @@
-import { TeamDataTableForm } from "@/components/ui/Users/team-data-table-form";
-
-export default function CreateTeam() {
-  return <TeamDataTableForm />;
-}
